ETL code refactorization
